drivers: Add light API and a driver for RGB-CT light
This adds a light API to control light device.
This is quite similar to the existing led API but this
could address some specific needs of light device in a
way simpler than with led API.
This also adds a first driver for RGB-CT light devices.
And test it quickly, this also introduce a new a shell command.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>