ESPHome: Make passeword optionnal
The password is deprecated in ESPHome and we don't correctly
handle error related to password in the API component.
Make password optionnal (which is aligned with EPSHome).

Note:
Even when the password is set, there are no encryption.
This is another ESPHome protocol and feature to implement.
The main purpose of password is to make sure that we could
only add a device to homeassistant if we know its password.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>